refactor: #83 enum 파일 분리
